feat: rename email-grantees to send-email and redesign with two-column layout
Rename route, files, and components from email-grantees to send-email.
Redesign page with a two-column layout: filters on the left, email
composer on the right with recipients inline as a "To" row above the
subject line.
