+++ app/community/[communityId]/manage/send-email/loading.tsx
@@ -4,7 +4,7 @@ export default function Loading() {
   return (
     <output
       aria-live="polite"
-      aria-label="Loading email grantees"
+      aria-label="Loading send email"
       className="flex items-center justify-center min-h-screen"
     >
       <Spinner />
